Added Farm Animal Tracking Code
This application tracks animals by RFID chip attached to the animal.  The hardware needed is a Raspberry PI PICO W, an OLED Display (4 pin, 128,64px) and an RFID  RC522 board.

This is a working package.  I consider it 90% complete only because I need to build the functionality to assign an animal's name to the card ID and display it on the screen with the card number.