Another Image
I found that the existing image did not work well with Ogilvie School's Chrome books.  They have a few different profiles; the original PICO image didn't work on all of them.  This one seemed to work on all of them.  However, this is just a hunch.  It could have been the chip version or because we were working with Pico Wi-Fi boards.

When 20-something 4th to 8th graders are running around the classroom, chaos could have been the issue.